+++ test/support/test_filter.ex
@@ -1,5 +1,5 @@
 defmodule Sentry.TestFilter do
-  @behaviour Sentry.Filter
+  @behaviour Sentry.EventFilter
 
   def exclude_exception?(%ArithmeticError{}, :plug), do: true
   def exclude_exception?(_, _), do: false