+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## master
 
 * Enhancements
-  * Allow filtering which exceptions are sent via `Sentry.Filter` behaviour
+  * Allow filtering which exceptions are sent via `Sentry.EventFilter` behaviour
 
 * Bug Fixes
   * Fix usage of deprecated modules